Refactored for Chart to be single component. Style updates.
